devtools/client/shared/components/test/chrome/test_notification_box_05.html


<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test for Notification Box. The test is checking:
* the close button is not present when displayCloseButton is false
-->
<head>
  <meta charset="utf-8">
  <title>Notification Box</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">

'use strict'


window.onload = async function () {
  try {
    const ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    const React = browserRequire("devtools/client/shared/vendor/react");
    const { NotificationBox, PriorityLevels } = browserRequire("devtools/client/shared/components/NotificationBox");

    // Test rendering with close button disabled
    const boxElement = React.createElement(NotificationBox, {displayCloseButton: false});
    const notificationBox = TestUtils.renderIntoDocument(boxElement);
    const notificationNode = ReactDOM.findDOMNode(notificationBox);



    // Append a notification.
    notificationBox.appendNotification(
      "Info message",
      "id1",
      null,
      PriorityLevels.PRIORITY_INFO_LOW,
      [],
      (e) => false,
    );

    // Ensure close button is not present
    const linkNode = notificationNode.querySelector(
      ".messageCloseButton");

    ok(!linkNode, "Close button is not present");

  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
};
</script>
</pre>
</body>
</html>
